requirements.txt
dvc[gdrive]
numpy<2
torch<2.4.0
fastapi
pandas
pyyaml
mlflow
apache-airflow[cncf.kubernetes]
argparse
Flask
flask-session2
psycopg2-binary
torchtext==0.13.0
ipywidgets
tqdm
nltk
onnx
onnxscript
gradio
email-validator>=2.0
waitress
argparse
pybind11>=2.12